refactor main package -> editor (forgot to do that earlier)
