+++ app/src/main/java/com/study/thesuperiorstanislav/edaapp/data/source/CircuitRepository.kt
@@ -1,7 +1,7 @@
 package com.study.thesuperiorstanislav.edaapp.data.source
 
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Circuit
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.draw.DrawObject
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.Circuit
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw.DrawObject
 
 object CircuitRepository: CircuitDataSource {
