+++ app/src/main/java/com/study/thesuperiorstanislav/edaapp/editor/EditorContract.kt
@@ -1,10 +1,10 @@
-package com.study.thesuperiorstanislav.edaapp.main
+package com.study.thesuperiorstanislav.edaapp.editor
 
 import com.study.thesuperiorstanislav.edaapp.BasePresenter
 import com.study.thesuperiorstanislav.edaapp.BaseView
 import com.study.thesuperiorstanislav.edaapp.UseCase
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Circuit
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.draw.DrawObject
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.Circuit
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw.DrawObject
 
 interface EditorContract {
