+++ app/src/main/java/com/study/thesuperiorstanislav/edaapp/editor/domain/model/draw/DrawObject.kt
@@ -1,3 +1,3 @@
-package com.study.thesuperiorstanislav.edaapp.main.domain.model.draw
+package com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw
 
 data class DrawObject(var drawPoint: DrawPoint,var objectType: ObjectType,var drawType: DrawType)
