+++ app/src/main/java/com/study/thesuperiorstanislav/edaapp/editor/domain/model/draw/DrawPoint.kt
@@ -1,0 +1,3 @@
+package com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw
+
+data class DrawPoint (var x: Float, var y: Float)
