+++ app/src/main/java/com/study/thesuperiorstanislav/edaapp/editor/domain/model/draw/ObjectType.kt
@@ -1,0 +1,7 @@
+package com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw
+
+enum class ObjectType {
+    Net,
+    Pin,
+    Connector
+}
