--- app/src/main/java/com/study/thesuperiorstanislav/edaapp/main/domain/model/draw/DrawPoint.kt
@@ -1,3 +1,0 @@
-package com.study.thesuperiorstanislav.edaapp.main.domain.model.draw
-
-data class DrawPoint (var x: Float, var y: Float)
