--- app/src/main/java/com/study/thesuperiorstanislav/edaapp/main/domain/model/draw/ObjectType.kt
@@ -1,7 +1,0 @@
-package com.study.thesuperiorstanislav.edaapp.main.domain.model.draw
-
-enum class ObjectType {
-    Net,
-    Pin,
-    Connector
-}
