+++ app/src/main/java/com/study/thesuperiorstanislav/edaapp/utils/graphics/Placer.kt
@@ -1,14 +1,14 @@
 package com.study.thesuperiorstanislav.edaapp.utils.graphics
 
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Circuit
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Element
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Element.DrawType.*
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Net
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.Point
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.draw.DrawObject
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.draw.DrawPoint
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.draw.DrawType
-import com.study.thesuperiorstanislav.edaapp.main.domain.model.draw.ObjectType
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.Circuit
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.Element
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.Element.DrawType.*
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.Net
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.Point
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw.DrawObject
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw.DrawPoint
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw.DrawType
+import com.study.thesuperiorstanislav.edaapp.editor.domain.model.draw.ObjectType
 import org.jetbrains.anko.collections.forEachReversedWithIndex
 
 class Placer(private val drawMatrix: Array<Array<DrawObject?>>,
